agency_fund_assignment/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
google-generativeai==0.3.2
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
